# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pessoas",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Mostra uma pessoa",
                "description": "Mostra uma pessoa",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Pessoa retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a pessoa."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Atualização de pessoa",
                "description": "update",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nomePessoa",
                                    "dataNascimento",
                                    "cpf",
                                    "rg"
                                ],
                                "properties": {
                                    "nomePessoa": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Leonardo"
                                    },
                                    "dataNascimento": {
                                        "description": "Data de nascimento",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF",
                                        "type": "string"
                                    },
                                    "rg": {
                                        "schema": "string",
                                        "description": "RG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Atualização realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a atualização."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Exclusão de pessoa",
                "description": "destroy",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Exclusão realizada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível realizar a exclusão."
                    }
                }
            }
        },
        "/api/v1/telefonePessoas": {
            "get": {
                "tags": [
                    "TelefonePessoas"
                ],
                "summary": "Lista de telefones de pessoas",
                "description": "Retorna uma lista de telefones de pessoas",
                "operationId": "getTelefonePessoasList",
                "responses": {
                    "200": {
                        "description": "Sucesso. Lista de telefones de pessoas retornada com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível retornar a lista de telefones de pessoas."
                    }
                }
            },
            "post": {
                "tags": [
                    "TelefonePessoas"
                ],
                "summary": "Cadastro de telefone de pessoa",
                "description": "store",
                "operationId": "c63d2bc2616dcf908ca356489c358ed1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telefonePessoa": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "idPessoa": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Telefone de pessoa cadastrado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível cadastrar o telefone de pessoa."
                    }
                }
            }
        },
        "/api/v1/telefonePessoas/{id}": {
            "get": {
                "tags": [
                    "TelefonePessoas"
                ],
                "summary": "Exibir telefone de pessoa",
                "description": "show",
                "operationId": "dd92c920d2ef07ec5beb21a7a179d8a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Telefone de pessoa exibido com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível exibir o telefone de pessoa."
                    }
                }
            },
            "put": {
                "tags": [
                    "TelefonePessoas"
                ],
                "summary": "Atualizar telefone de pessoa",
                "description": "update",
                "operationId": "0b86265f67c0f684da476c4ee0cbf088",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telefonePessoa": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "idPessoa": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso. Telefone de pessoa atualizado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível atualizar o telefone de pessoa."
                    }
                }
            },
            "delete": {
                "tags": [
                    "TelefonePessoas"
                ],
                "summary": "Deletar telefone de pessoa",
                "description": "destroy",
                "operationId": "6f1d2c19ee32142ba9d60fc9c1b6a8e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone de pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso. Telefone de pessoa deletado com sucesso."
                    },
                    "400": {
                        "description": "Erro. Não foi possível deletar o telefone de pessoa."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "scheme": "http",
                "name": "bearer",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}